add GPIO control for run levels
Please note that driver performs no sanity-check for GPIO
state changes. It simply writes new RUN level by first
writing the GPIO1 and then the GPIO2. So for example, if
original level is 0 and level 3 is requested as new level,
the transition done by driver will be:

0 => 1 => 3. If preferred transition is 0 => 2 => 3, then
it is driver users responcebility to explicitly request
transition from level 0 to level 2 and then to 3.