gpio: bd71828: Initial support for ROHM BD71828 PMIC GPIOs
ROHM BD71828 PMIC contains 3 pins which can be configured by OTP
to be used as general purpose outputs. Allow them to be controlled
via GPIO framework.

The driver assumes all of the pins are configured as GPIO outputs
and trusts that the reserved pins in other OTP configurations are
excluded from control using "gpio-reserved-ranges"