misc: bd71828: Support LID position interrupt
The BD71828 has a PIN which is dedicated to be used for a HALL sensor.
When the state of this pin is toggled by the HALL sensor, the PMIC will
generate a specific IRQ. The intended use-case for this HALL IRQ is
detecting closing the lid.

Support sending the LID event to user-space when this IRQ is triggered.

This commit squashes bunch of commits cherry-picked from the old
BD71828 development branch.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>