regulator: bd71828: remove unnecessary DT property reading
Original (bad) idea was to describe bucks which are bound to the
run-level control mechanism in one array located in MFD DT node.

Later I noticed that it is nicer to put this information into
regulator node for the specific regulator that is controlled by
run-level. This makes DT much more clear.

Old code which reads the array of buck was left in regulator
driver - and it also used wrong DT property name. Remove this
unnecessary reading.