regulator: bd71828: Restrict run-level setting/getting.
Restrict run-level setting/geting to be allowed only via pointer
to regulator which is controlled by run-levels. Do this in order
to not need to align the 'GPIO/I2C' control information.

Other optuion would be to only present the I2C or GPIO control
information in one globally accessible struct. Let's not go to
that route for now.