power: supply: bd71828: Fix the measured current value
BD71828 can report charging current. According to the data-sheet,
when register value increases by 1, current is increased by 0.33 mA
(when RSENS is 30mOhm)

Old BD71827 code did assume register value is directly the current
in mA and did not apply any scaling. I don't know if this is a
bug on BD71827 - but on BD71828 this yields bad current values to
be reported via sysfs. Fix this by using register value to uA
conversion factor 330 on BD71828. I kept BD71827 conversion factor
as 1000 - let's change it only if issues are reported or when it is
clarified if we can get the RSENS from device tree and convert it
nicely to a proper scaling factor. 330 can be used as default for
BD71828.