regulator: bd96801: ROHM BD96801 PMIC regulators
The ROHM BD96801 "Scalable PMIC" is an automotive grade PMIC which can
scale to different applications by allowing chaining of PMICs. The PMIC
also supports various protection features which can be configured either
to fire IRQs - or to shut down power outputs when failure is detected.

The driver supports the basic control of BUCKs and LDOs and configuring
Over/Under-voltage, over-current, over-temperature protections via
device-tree. Following constrains are worth noting:

The voltages can be enabled/disabled and protections configured only
when the PMIC is in STANDBY state.

The protections can't be disabled. UVP limit can't be configured.
OVP and OCP limits can be configured.

OVD and UVD for a regulator have common limit.
OCP and thermal protections are implemented by turning all INTB
notifications fatal. This means that OVD/UVD can not be used together
with OCP and thermal protection.

Thermal limits can't be configured. TSD is fixed to 175 Celsius, and
thermal warning is fixed close to 140 Celsius.

LDO's do not have own temperature monitor. The LDO limits use PMIC
core temperature.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Revision history:
RFCv1 => RFCv2:
- Use devm_kcalloc() instead of devm_kzalloc() where appropriate
- Use devm_kmemdup()
- Add MODULE_DEVICE_TABLE
- drop MODULE_ALIAS

This patch requires the regulator_irq_helper() IRQ name duplication
patch which is already in regulator tree 'for-6.9' -branch.
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git/commit/?h=for-6.9&id=7ab681ddedd4b6dd2b047c74af95221c5f827e1d

Note. This patch implements controls which can only be done when the
BD96801 PMIC is in a 'standby' state. These controls include the
protection settings and 'initial' voltage settings.

The PMIC has a STBY pin which is used to control the PMIC power-state.

I don't have visibility to the design process or to most of the customer
use cases - so following is speculation:
 - Original idea seems to have been that, when the PMIC goes to STBY
   state, controlling processor will be powered-off. In the original DS1
   version all power outputs from the PMIC were turned off when PMIC
   entered STBY.
 - Someone somewhere needed to be able to keep the controlling processor
   and some peripherials powered on when the PMIC is in STBY. A very
   potential reason is to be able to do the initial voltage / protection
   configurations by software. A configuration option to keep desired
   power outputs enabled when PMIC is in STBY was added.
 - Current PMIC state can be read from the PMIC registers.

Obvious(?) dilemma for these 'STBY only' configurations is that it is
difficult to design generic way for the PMIC driver to synchronize the
configurations with the state changes. Designing it would at least require
knowledge on:
 - What is the PMIC STBY-state used for? Is it intended to be a
   power-off (or other low-power) state for the system? Or, is it
   intended to be used just a state where some of the configurations are
   done?
 - How is the hardware designed to control the STBY pin?

We *could* assume the STBY pin is controlled by a GPIO. We could grab
the GPIO here, and export functions to toggle the STBY state. We could
then 'lock' the state for the duration of config changes. But problem
is, I don't know what is the right thing to do.

So, this version of the patch just checks the PMIC state - and attempts
to write the configs if PMIC is in STBY state. This is racy unless the
entity controlling the STBY-pin is somehow (external to this driver)
synchronized. We could send a notification when configurations are
completed to help building such synchronization but this is not done as
I don't have a real-world user for it.

I am unsure if these 'STBY-only' controls should be just dropped, or if
they should be kept here to serve as an example / basis to build on. If
they are removed, it will be very hard for anyone using this PMIC to
implement those controls. OTOH, if they are kept here, it may
potentially encourage 'works on my machine' designs - although, I
believe that those who really need to do 'STBY-only' configs at run-time
will find a way to toggle the STBY-pin in a way that is synchronized
with the configs here.