mfd: BD96801: Support IRQ setup for DS2
IRQ setup will be fixed for DS2. The ERRB and INTB related
information will be placed in separate registers in order to
avoid race conditions.

Let's handle the INTB irqs using standard regmap-irq irq chip.
ERRB interrutps are expected to be resetting the SoC so let's
ignore them for now.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>