dt-bindings: mfd: regulator: Drop BD70528 support
The only known BD70528 use-cases are such that the PMIC is controlled
from separate MCU which is not running Linux. I am not aware of
any Linux driver users. Furthermore, it seems there is no demand for
this IC. Let's ease the maintenance burden and drop the driver. We can
always add it back if there is sudden need for it.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>

---
Bah. My heart is bleeding as I personally liked playing with this IC.
Still, It makes no sense to waste the time of you guys by maintaining
something which is not likely to be used. Please let me know if some
of you think the binding is needed.