regulator: irq_helpers: Allow omitting map_event for simple IRQs
If the device has dedicated IRQ (or irq controller already splits the IRQ
appropriately) for single error indication on a single regulator then the
map-event callback has not much to do.

Simplify the usage for such devices by using a common helper function to
return the regulator and the reason when map_event is not populated.

Signed-off-by: Matti Vaittinen <matti.vaittinen@fi.rohmeurope.com>

---
Internal development history to track changes
 - regulator: irq_helpers: Fix single_bit_set