dt-bindings: Fix maintainer email for a few ROHM ICs
The email backend used by ROHM keeps labeling patches as spam. This can
result to missing the patches.

Switch my mail address from a company mail to a personal one.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---

I did this a while ago for my email addresses at the MAINTAINERS. Forgot
the dt-bindings and doing the conversion for bindings now.