dts: bbb: add BD71885 and BD2659 PMICs to bbb dts for testing
Add the PMIC nodes to beagle-bone black dts to enable testing the pmic
drivers on beagle-bone black. The PMICs should be connected to the I2C2.

THIS COMMIT IS NEVER INTENDED TO BE USED FOR ANYTHING BUT TESTING OR
PERHAPS AS AN EXAMPLE OF A DTS BINDING.

THIS COMMIT SHOULD NEVER GO TO UPSTREAM.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
If you see this in a mail list then something somewhere has gone
terribly wrong and I do apologize :)