power: regulator: bd2659: Fix the regulator ranges
The macro initializing the voltage ranges for regulators expected an
array of ranges as argument. The macro ARRAY_SIZE() was then used to
initialize the amount of ranges. The calling code however passed a
pointer to first array element as argument, thus leaving the amount of
the ranges to zero resulting failures to convert register values to
voltages or voltages to register values.

Fix this by passing the actual array as argument to the macro.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Fixes: 4af38b27b93f ("regulator: Support ROHM BD2659")