net-next: mvpp2: relax return value check for IRQ get
fwnode_irq_get[_byname]() were changed to not return 0 anymore.

Drop check for return value 0.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Revision history:
v4 = v5:
Fix the subject, mb1232 => mvpp2

The first patch of the series changes the fwnode_irq_get() so this depends
on the first patch of the series and should not be applied alone.