pinctrl: wpcm450: relax return value check for IRQ get
fwnode_irq_get[_byname]() were changed to not return 0 anymore. The
special error case where device-tree based IRQ mapping fails can't no
longer be reliably detected from this return value. This yields a
functional change in the driver where the mapping failure is treated as
an error.

The mapping failure can occur for example when the device-tree IRQ
information translation call-back(s) (xlate) fail, IRQ domain is not
found, IRQ type conflicts, etc. In most cases this indicates an error in
the device-tree and special handling is not really required.

One more thing to note is that ACPI APIs do not return zero for any
failures so this special handling did only apply on device-tree based
systems.

Drop the special (no error, just skip the IRQ) handling for DT mapping
failures as these can no longer be separated from other errors at driver
side.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
Reviewed-by: Jonathan Neuschäfer <j.neuschaefer@gmx.net>
Acked-by: Linus Walleij <linus.walleij@linaro.org>

---
Revision history:
v4 => v5:
Fix typo in subject "elax" => "relax"

Please note, I took Linus' reply to v4 cover-letter as ack && added the
tag. Please let me know if this was not Ok.

The special handling in this driver was added when fixing a problem
where returning zero from fwnode_irq_get[_byname]() was treated as
succes yielding zero being used as a valid IRQ by the driver.
f4a31facfa80 ("pinctrl: wpcm450: Correct the fwnode_irq_get() return value check")
The commit message does not mention if choosing not to abort the probe
on device-tree mapping failure (as is done on other errors) was chosen
because: a) Abort would have broken some existing setup. b) Because skipping
an IRQ on failure is "the right thing to do", or c) because it sounded like
a way to minimize risk of breaking something.

If the reason is a) - then I'd appreciate receiving some more
information and a suggestion how to proceed (if possible). If the reason
is b), then it might be best to just skip the IRQ instead of aborting
the probe for all errors on IRQ getting. Finally, in case of c), well,
by acking this change you will now accept the risk :)

The first patch of the series changes the fwnode_irq_get() so this depends
on the first patch of the series and should not be applied alone.